Reorganize the code to make it more readable
