+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "functions.h": "c"
+        "functions.h": "c",
+        "windowlayer.h": "c"
     }
 }
