+++ inc/windowLayer.h
@@ -1,0 +1,16 @@
+#ifndef _WINDOWLAYER_H
+#define _WINDOWLAYER_H
+
+#include "functions.h"
+
+extern UINT8 swap;
+
+extern unsigned char windowmap[];
+
+extern void updateCoinsCounter();
+
+extern void updateHealth();
+
+extern void updateScore();
+
+#endif
